move send screen logic + state to hook
