+++ packages/renderer/src/renderer/screens/Home/Send/index.tsx
@@ -7,16 +7,13 @@ import * as FA from '@fortawesome/free-solid-svg-icons';
 import * as themeVariables from '../../../theme/variables';
 import { FlexColumn, FlexRow } from '@polkadot-live/ui/styles';
 
-import { Config as ConfigRenderer } from '@ren/config/processes/renderer';
-import { chainCurrency, chainUnits } from '@ren/config/chains';
+import { chainCurrency } from '@ren/config/chains';
 import { Identicon, MainHeading } from '@polkadot-live/ui/components';
 import { useConnections } from '@app/contexts/common/Connections';
-import { useEffect, useRef, useState } from 'react';
-import { ellipsisFn, unitToPlanck } from '@w3ux/utils';
-import { getAddressChainId } from '@app/Utils';
+import { useState } from 'react';
+import { ellipsisFn } from '@w3ux/utils';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { ActionButton, InputWrapper } from './Wrappers';
-import { getSpendableBalance } from '@ren/utils/AccountUtils';
 import { getBalanceText } from '@ren/utils/TextUtils';
 import {
   AccountNameWithTooltip,
@@ -29,45 +26,37 @@ import {
   TriggerContent,
 } from './SendHelpers';
 
-import type {
-  AccountSource,
-  LedgerLocalAddress,
-  LocalAddress,
-} from '@polkadot-live/types/accounts';
-import type {
-  ActionMeta,
-  ExTransferKeepAliveData,
-} from '@polkadot-live/types/tx';
-import type BigNumber from 'bignumber.js';
-import type { ChainID } from '@polkadot-live/types/chains';
-import type { ChangeEvent } from 'react';
-import type { SendAccordionValue, SendRecipient } from './types';
+import type { SendAccordionValue } from './types';
 import { DialogRecipient } from './Dialogs';
-
-const TOKEN_TRANSFER_LIMIT = 100;
+import { useSendNative } from '@ren/renderer/hooks/useSendNative';
 
 export const Send: React.FC = () => {
+  const {
+    fetchingSpendable,
+    progress,
+    receiver,
+    sendAmount,
+    sender,
+    senderNetwork,
+    spendable,
+    summaryComplete,
+    validAmount,
+    getReceiverAccounts,
+    getSenderAccounts,
+    getSenderAccountName,
+    handleProceedClick,
+    handleResetClick,
+    handleSendAmountBlur,
+    handleSendAmountChange,
+    handleSendAmountFocus,
+    handleSenderChange,
+    proceedDisabled,
+    setReceiver,
+  } = useSendNative();
+
   /**
    * Addresses fetched from main process.
    */
-  const [addressMap, setAddressMap] = useState(
-    new Map<AccountSource, (LocalAddress | LedgerLocalAddress)[]>()
-  );
-  const addressMapRef = useRef<typeof addressMap>(addressMap);
-  const [updateCache, setUpdateCache] = useState(false);
-  const [progress, setProgress] = useState(0);
-
-  const [sender, setSender] = useState<null | string>(null);
-  const [receiver, setReceiver] = useState<null | SendRecipient>(null);
-  const [senderNetwork, setSenderNetwork] = useState<ChainID | null>(null);
-  const [sendAmount, setSendAmount] = useState<string>('0');
-
-  const [fetchingSpendable, setFetchingSpendable] = useState(false);
-  const [spendable, setSpendable] = useState<BigNumber | null>(null);
-  const [validAmount, setValidAmount] = useState(true);
-
-  const [summaryComplete, setSummaryComplete] = useState(false);
-
   const { darkMode } = useConnections();
   const theme = darkMode ? themeVariables.darkTheme : themeVariables.lightThene;
 
@@ -77,371 +66,6 @@ export const Send: React.FC = () => {
   const [accordionValue, setAccordionValue] = useState<SendAccordionValue[]>([
     'section-sender',
   ]);
-
-  /**
-   * Fetch stored addresss from main when component loads.
-   */
-  useEffect(() => {
-    const fetch = async () => {
-      const serialized = (await window.myAPI.rawAccountTask({
-        action: 'raw-account:getAll',
-        data: null,
-      })) as string;
-
-      const parsedMap = new Map<AccountSource, string>(JSON.parse(serialized));
-
-      for (const [source, ser] of parsedMap.entries()) {
-        switch (source) {
-          case 'vault':
-          case 'read-only':
-          case 'wallet-connect': {
-            const parsed: LocalAddress[] = JSON.parse(ser);
-            addressMapRef.current.set(source, parsed);
-            setUpdateCache(true);
-            break;
-          }
-          case 'ledger': {
-            const parsed: LedgerLocalAddress[] = JSON.parse(ser);
-            addressMapRef.current.set(source, parsed);
-            setUpdateCache(true);
-            break;
-          }
-          default: {
-            continue;
-          }
-        }
-      }
-    };
-
-    fetch();
-  }, []);
-
-  /**
-   * Mechanism for updating address map state from an async process.
-   */
-  useEffect(() => {
-    if (updateCache) {
-      setAddressMap(addressMapRef.current);
-      setUpdateCache(false);
-    }
-  }, [updateCache]);
-
-  /**
-   * Control summary complete flag.
-   */
-  useEffect(() => {
-    setSummaryComplete(false);
-  }, [sender, receiver, sendAmount]);
-
-  /**
-   * Progress bar controller.
-   */
-  useEffect(() => {
-    let conditions = 0;
-
-    sender && (conditions += 1);
-    receiver && (conditions += 1);
-    sendAmount !== '0' && sendAmount !== '' && validAmount && (conditions += 1);
-
-    switch (conditions) {
-      case 1: {
-        setProgress(100 / 3);
-        break;
-      }
-      case 2: {
-        setProgress((100 / 3) * 2);
-        break;
-      }
-      case 3: {
-        setProgress(100);
-        break;
-      }
-      default: {
-        setProgress(0);
-        break;
-      }
-    }
-  }, [sender, receiver, sendAmount]);
-
-  /**
-   * Handle proceed click.
-   */
-  const handleProceedClick = async () => {
-    if (!(senderNetwork && sender && receiver)) {
-      return;
-    }
-
-    // NOTE: Disable Polkadot transfers in alpha releases.
-    if (senderNetwork === 'Polkadot') {
-      return;
-    }
-
-    setSummaryComplete(true);
-
-    // Data for action meta.
-    const senderObj = getSenderAccounts().find(
-      ({ address }) => address === sender
-    )!;
-
-    const recipientObj = {
-      address: receiver.address,
-      accountName:
-        receiver.managed && receiver.accountName
-          ? receiver.accountName
-          : ellipsisFn(receiver.address, 8),
-    };
-
-    const sendAmountPlanck: string = unitToPlanck(
-      sendAmount.toString(),
-      chainUnits(senderNetwork)
-    ).toString();
-
-    // Specific data for transfer extrinsic.
-    const balanceData: ExTransferKeepAliveData = {
-      recipientAddress: recipientObj.address,
-      recipientAccountName: recipientObj.accountName,
-      sendAmount: sendAmountPlanck,
-    };
-
-    // Action meta.
-    const actionMeta: ActionMeta = {
-      accountName: senderObj.name,
-      source: senderObj.source,
-      action: 'balances_transferKeepAlive',
-      from: sender,
-      pallet: 'balances',
-      method: 'transferKeepAlive',
-      chainId: senderNetwork,
-      data: JSON.stringify(balanceData),
-      args: [recipientObj.address, sendAmountPlanck],
-    };
-
-    // Send extrinsic to action window.
-    window.myAPI.relayModeFlag('isBuildingExtrinsic', true);
-    const extrinsicsViewOpen = await window.myAPI.isViewOpen('action');
-
-    if (!extrinsicsViewOpen) {
-      // Relay init task to extrinsics window after its DOM has loaded.
-      window.myAPI.openWindow('action', {
-        windowId: 'action',
-        task: 'action:init',
-        serData: JSON.stringify(actionMeta),
-      });
-
-      // Analytics.
-      window.myAPI.umamiEvent('window-open-extrinsics', {
-        action: `send-transfer-keep-alive`,
-      });
-    } else {
-      window.myAPI.openWindow('action');
-
-      // Send init task directly to extrinsics window if it's already open.
-      ConfigRenderer.portToAction?.postMessage({
-        task: 'action:init',
-        data: JSON.stringify(actionMeta),
-      });
-    }
-  };
-
-  /**
-   * Handle clicking the reset button.
-   */
-  const handleResetClick = () => {
-    setSender(null);
-    setSenderNetwork(null);
-    setReceiver(null);
-    setSpendable(null);
-    setValidAmount(true);
-    setSendAmount('0');
-    setAccordionValue(['section-sender']);
-  };
-
-  /**
-   * Sender value changed callback.
-   */
-  const handleSenderChange = async (senderAddress: string) => {
-    setFetchingSpendable(true);
-
-    const chainId = getAddressChainId(senderAddress);
-    setSender(senderAddress);
-    setSenderNetwork(chainId);
-
-    const result = await getSpendableBalance(senderAddress, chainId);
-    setSpendable(result);
-    setFetchingSpendable(false);
-
-    // Reset other send fields.
-    setReceiver(null);
-    setSendAmount('0');
-    setValidAmount(true);
-  };
-
-  /**
-   * Utility to truncate a send amount to the network's allowable decimal places.
-   */
-  const truncateDecimals = (amount: string, chainId: ChainID): string => {
-    const decimals = chainUnits(chainId);
-    const [integerPart, decimalPart] = amount.split('.');
-
-    if (!decimalPart) {
-      return integerPart;
-    }
-
-    const truncatedDecimal = decimalPart.slice(0, decimals);
-    return `${integerPart}.${truncatedDecimal}`;
-  };
-
-  /**
-   * Removes leading zeros from a numeric string but keeps a single zero if a decimal follows.
-   */
-  const removeLeadingZeros = (value: string): string => {
-    // Remove unnecessary leading zeros.
-    let cleaned = value.replace(/^0+(?=\d)/, '');
-
-    // If the first character is ".", prepend "0"
-    if (cleaned.startsWith('.')) {
-      cleaned = '0' + cleaned;
-    }
-
-    // Ensure empty string returns "0"
-    return cleaned || '0';
-  };
-
-  /**
-   * Send amount changed callback.
-   */
-  const handleSendAmountChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const val = event.target.value;
-
-    // Check for sender network and spendable flag.
-    if (!(senderNetwork !== null && spendable !== null)) {
-      setValidAmount(false);
-      return;
-    }
-    // Check for zero values.
-    if (val === '' || val === '0') {
-      setSendAmount(val);
-      setValidAmount(true);
-      return;
-    }
-    if (!isNaN(Number(val))) {
-      // Truncate to network's allowable decimal places and remove any leading zeros.
-      const tmp: string = removeLeadingZeros(val);
-      const amount = truncateDecimals(tmp, senderNetwork);
-      setSendAmount(amount);
-
-      // Check for negative value.
-      if (Number(amount) < 0) {
-        setSendAmount(amount);
-        setValidAmount(false);
-        return;
-      }
-
-      // NOTE: Limit send amount to 100 tokens in alpha releases.
-      if (Number(amount) > TOKEN_TRANSFER_LIMIT) {
-        setSendAmount(amount);
-        setValidAmount(false);
-        return;
-      }
-
-      // Check if send amount is less than spendable amount.
-      const units = chainUnits(senderNetwork);
-      const bnAmountAsPlanck = unitToPlanck(amount, units);
-      setValidAmount(spendable.gte(bnAmountAsPlanck));
-      return;
-    }
-
-    setValidAmount(false);
-  };
-
-  /**
-   * Remove zero when send amount field focused.
-   */
-  const handleSendAmountFocus = () => {
-    if (sendAmount === '0') {
-      setSendAmount('');
-    }
-  };
-
-  /**
-   * Add zero when send amount field blurred and empty.
-   */
-  const handleSendAmountBlur = () => {
-    if (sendAmount === '') {
-      setSendAmount('0');
-    }
-  };
-
-  /**
-   * Handle copy to clickboard.
-   */
-  const handleClipboardCopy = async (text: string) =>
-    await window.myAPI.copyToClipboard(text);
-
-  /**
-   * Return all addresses capable of signing extrinsics.
-   */
-  const getSenderAccounts = () => {
-    const targetSources: AccountSource[] = ['vault', 'wallet-connect'];
-    let result: LocalAddress[] = [];
-
-    for (const source of targetSources) {
-      const addresses = addressMap.get(source);
-      if (!addresses || addresses.length === 0) {
-        continue;
-      }
-
-      // NOTE: Disable Polkadot transfers in alpha releases.
-      const filtered = (addresses as LocalAddress[]).filter(
-        ({ address }) => getAddressChainId(address) !== 'Polkadot'
-      );
-
-      result = result.concat(filtered);
-    }
-    return result.sort((a, b) => a.name.localeCompare(b.name));
-  };
-
-  /**
-   * Return all addresses for receiver address list.
-   */
-  const getReceiverAccounts = () => {
-    let result: (LocalAddress | LedgerLocalAddress)[] = [];
-    for (const addresses of addressMap.values()) {
-      result = result.concat(addresses);
-    }
-
-    // Filter accounts on sender address network.
-    return (
-      result
-        .filter(({ address }) => {
-          if (!senderNetwork) {
-            return true;
-          } else {
-            return getAddressChainId(address) === senderNetwork;
-          }
-        })
-        // NOTE: Disable Polkadot transfers in alpha releases.
-        .filter(({ address }) => getAddressChainId(address) !== 'Polkadot')
-        // Don't include sender in list.
-        .filter(({ address }) => address !== sender)
-        .sort((a, b) => a.name.localeCompare(b.name))
-    );
-  };
-
-  /**
-   * Conditions to enable the proceed button.
-   */
-  const proceedDisabled = () =>
-    sender === null ||
-    receiver === null ||
-    sendAmount === '0' ||
-    sendAmount === '' ||
-    // NOTE: Limit token transfers to 100 tokens in alpha releases.
-    (!isNaN(Number(sendAmount)) && Number(sendAmount) > TOKEN_TRANSFER_LIMIT) ||
-    // NOTE: Disable Polkadot transfers in alpha releases.
-    senderNetwork === 'Polkadot' ||
-    !validAmount ||
-    summaryComplete;
 
   /**
    * Handle clicking a green next step arrow.
@@ -464,13 +88,10 @@ export const Send: React.FC = () => {
   };
 
   /**
-   * Utility for getting sender and receiver account names.
+   * Handle copy to clickboard.
    */
-  const getSenderAccountName = () =>
-    !sender
-      ? '-'
-      : getSenderAccounts().find(({ address }) => address === sender)?.name ||
-        ellipsisFn(sender, 12);
+  const handleClipboardCopy = async (text: string) =>
+    await window.myAPI.copyToClipboard(text);
 
   const emptySenders = getSenderAccounts().length === 0;
   const emptyReceivers = getReceiverAccounts().length === 0;
@@ -754,7 +375,10 @@ export const Send: React.FC = () => {
                     <ActionButton
                       style={{ flex: 1 }}
                       $backgroundColor={'var(--button-background-secondary)'}
-                      onClick={() => handleResetClick()}
+                      onClick={() => {
+                        handleResetClick();
+                        setAccordionValue(['section-sender']);
+                      }}
                       disabled={false}
                     >
                       <FontAwesomeIcon
